Modularize zinc a bit more before introducing additional dependencies.
