+++ src/scala/org/pantsbuild/zinc/options/BUILD
@@ -1,16 +1,13 @@
 scala_library(
   provides=scala_artifact(
     org='org.pantsbuild',
-    name='zinc-cache',
+    name='zinc-options',
     repo=public,
     publication_metadata=pants_library('The SBT incremental compiler for nailgun')
   ),
   dependencies=[
-    '3rdparty:guava',
-    '3rdparty:jsr305',
-  ],
-  exports=[
-    '3rdparty:guava',
+    '3rdparty/jvm/org/scala-lang/modules:scala-java8-compat',
+    '3rdparty/jvm/org/scala-sbt:zinc',
   ],
   strict_deps=True,
   platform='java8',
