+++ src/scala/org/pantsbuild/zinc/options/Options.scala
@@ -2,7 +2,7 @@
  * Copyright (C) 2012 Typesafe, Inc. <http://www.typesafe.com>
  */
 
-package org.pantsbuild.zinc
+package org.pantsbuild.zinc.options
 
 import java.io.File
 import scala.annotation.tailrec
