+++ src/scala/org/pantsbuild/zinc/util/BUILD
@@ -1,16 +1,12 @@
 scala_library(
   provides=scala_artifact(
     org='org.pantsbuild',
-    name='zinc-cache',
+    name='zinc-util',
     repo=public,
     publication_metadata=pants_library('The SBT incremental compiler for nailgun')
   ),
   dependencies=[
-    '3rdparty:guava',
-    '3rdparty:jsr305',
-  ],
-  exports=[
-    '3rdparty:guava',
+    '3rdparty/jvm/org/scala-sbt:zinc',
   ],
   strict_deps=True,
   platform='java8',
